#1433 float -> double

+++ org.caleydo.view.pathway/src/org/caleydo/view/pathway/GLPathwayAugmentationRenderer.java
@@ -638,7 +638,7 @@ public class GLPathwayAugmentationRenderer {
 					float scaleConstant = 20;
 					if (mappingPerspective.getDataDomain().getTable() instanceof NumericalTable) {
 						NumericalTable table = (NumericalTable) mappingPerspective.getDataDomain().getTable();
-						dsMean = table.getDatasetStatistics().getMean();
+						dsMean = (float) table.getDatasetStatistics().getMean();
 						dsMean = (float) table.getNormalizedForRaw(table.getDefaultDataTransformation(), dsMean);
 					}
